fix: 페이지네이션 옵션 서비스레이어에서 레포지토리 레이어로 변경
